requirements.txt
python-telegram-bot
yfinance
mplfinance
pandas
numpy
matplotlib
python-dotenv
ta